[ADD] -fsanitize=address in src/Makefile which could detect memory issues at runtime. [ADD] Exclusive mutex protects the index of localized keys in remapIndex function. [ADD] The size of new_offset must be identical to the number of examples across partitions. [BUGFIX] Range<T>::evenDivide could produce the end of the last segment correctly. [MOD] Makes SArray<V>::setUnion serial temporarily. [MOD] SArray<V>::appendToFile could produce empty file if the range is empty. [ADD] SlotReader::appendToFile produces the third column (which is the element number) into partition infomation files. [MOD] BatchSolver initializes Valiation both for workers and servers. [MOD] FreqencyFilter<K>::queryKeys returns empty SArray if its internal CountMin is empty. [MOD] Ocean::dump returns true if received parameter_key is empty.

+++ src/base/shared_array.h
@@ -91,8 +91,8 @@ template<typename V> class SArray {
   // Element access:
   V back() const { CHECK(!empty()); return data_[size_-1]; }
   V front() const { CHECK(!empty()); return data_[0]; }
-  V& operator[] (int i) { return data_[i]; }
-  const V& operator[] (int i) const { return data_[i]; }
+  V& operator[] (size_t i) { return data_[i]; }
+  const V& operator[] (size_t i) const { return data_[i]; }
 
   // Modifiers
   void append(const SArray<V>& tail);
